% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_function.r
\name{cost_function}
\alias{cost_function}
\title{Utility function}
\usage{
cost_function(type, N_fleet, fleet, N_species, species)
}
\arguments{
\item{type}{type of cost function}

\item{fleet}{the fleet cost per event for every event in the fleet}

\item{species}{the species cost per event sampled (N)}

\item{N}{Number of events to apply the cost function to}
}
\description{
Utility function
}
\author{
Alistair Dunn & Johanna Pierre
}
\keyword{internal}
