% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMiterate.r
\name{EMiterate}
\alias{EMiterate}
\title{EM utility function}
\usage{
EMiterate(
  EM,
  objective.label,
  strata.label,
  sampling.fractions = NULL,
  quiet = TRUE,
  parallel = TRUE,
  cores = NULL
)
}
\arguments{
\item{EM}{an EM object}

\item{objective.label}{The label of the objective function to simulate and sample from}

\item{sampling.fractions}{Provides the sampling fractions for each strata. If NULL (the default), then determined using Neyman allocation}

\item{quiet}{If !quiet, then print information messages during the sampling}
}
\description{
This function simulates data
}
\author{
Alistair Dunn & Johanna Pierre
}
