% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMoptimise.r
\name{EMoptimise}
\alias{EMoptimise}
\title{EM utility function}
\usage{
EMoptimise(EM, EMiterations, sampling.fractions = NULL, stratum = NULL)
}
\arguments{
\item{EM}{An EM object}

\item{EMiterations}{the output of EMiternations}

\item{stratum}{Optionally, output the results from a specified stratum within strata.label}
}
\description{
This function summarises an EMiterate output into a simple table for viewing and plotting
}
\author{
Alistair Dunn & Johanna Pierre
}
