% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMoptimiseStrata.r
\name{EMoptimiseStrata}
\alias{EMoptimiseStrata}
\title{EM utility function}
\usage{
EMoptimiseStrata(
  EM,
  objective.labels,
  new.strata.label = "__new__",
  iter = 300,
  pops = 50
)
}
\arguments{
\item{EM}{an EM object}

\item{objective.labels}{The labels of the objective function to simulate and sample from (must be a vector of length > 1)}

\item{new.strata.label}{The label for the optimised strata calculated by the EMoptimiseStrata()}

\item{iter}{iter Maximum number of iterations (= generations) of the genetic algorithm (default is 300)}

\item{pops}{The dimension of each generations in terms of individuals (default is 50).}
}
\description{
This function uses the SamplingStrata package to find optimal strata for multiple objectives
}
\author{
Alistair Dunn & Johanna Pierre
}
