% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.table.subcommands.R
\name{get.table.subcommands}
\alias{get.table.subcommands}
\title{Utility function}
\usage{
get.table.subcommands()
}
\description{
Utility function
}
\author{
Alistair Dunn & Johanna Pierre
}
\keyword{internal}
