% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObject.r
\name{getObject}
\alias{getObject}
\title{Utility accessor function}
\usage{
getObject(EM, object, string)
}
\arguments{
\item{EM}{An EM object}

\item{object}{Character string with the command. If missing, returns a list of the commands in EM}

\item{string}{Character string with the command label. If missing, returns a list of labels of 'object' in EM}
}
\description{
Utility accessor function
}
\author{
Alistair Dunn & Johanna Pierre
}
