% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.config.file.R
\name{input.config.file}
\alias{input.config.file}
\title{Model configuration write function}
\usage{
input.config.file(
  filename,
  path = "",
  quiet = FALSE,
  NAs = c("NA", ".", "-", "-999"),
  rescale = TRUE
)
}
\arguments{
\item{path}{The path to the file (optional)}

\item{quiet}{Suppress print statements to screen (default is TRUE).}

\item{NAs}{Specify values used to represent missing or invalid values (i.e, NA, -999, etc), Defaults to c("NA", ".", "-", "-999")}

\item{rescale}{If true, then rescale the species distribtuins to have maximum=1}

\item{file}{The name of the input file containing EM model configuration}
}
\description{
This function reads a EM optimisation configuration file and returns a list object in R. Where each element is a command and subcommand from the EM definition file
}
\author{
Alistair Dunn & Johanna Pierre
}
