% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEMmap.r
\name{plotEMmap}
\alias{plotEMmap}
\title{Utility accessor function}
\usage{
plotEMmap(
  EM,
  type,
  label,
  plotit = TRUE,
  as.image = FALSE,
  zero.rm = TRUE,
  presence.only = FALSE,
  ...
)
}
\arguments{
\item{type}{Type of map (base or command)}

\item{label}{the command label}

\item{plotit}{either plot the data or return the object}

\item{as.image}{if true, convert character vectors to numeric, and plot as an image (default  = FALSE)}

\item{zero.rm}{Remove zeros and replace with NAs (default = TRUE)}

\item{presence.only}{Plot presense/absence (i.e., values > 0) only (default = FALSE)}

\item{...}{other arguments supplied to plot()}
}
\description{
Electronic Monitoring Video Review Rate Optimization. Pew Project: 2021-IF-02
}
\author{
Alistair Dunn & Johanna Pierre.
}
