% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEMsummary.r
\name{plotEMsummary}
\alias{plotEMsummary}
\title{Utility accessor function}
\usage{
plotEMsummary(data, cost = TRUE)
}
\arguments{
\item{data}{A summary iterate object from EMiterate}

\item{cost}{Plot the costs}
}
\description{
Electronic Monitoring Video Review Rate Optimization. Pew Project: 2021-IF-02
}
\author{
Alistair Dunn & Johanna Pierre.
}
