% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnegbin.r
\name{Rnegbin}
\alias{Rnegbin}
\title{Utility function}
\usage{
Rnegbin(n, mu = n, theta = stop("'theta' must be specified"))
}
\description{
Utility function
}
\author{
Alistair Dunn & Johanna Pierre
}
\keyword{internal}
